requirements.txt
flask>=3.0.3
sqlalchemy>=2.0.25
google-generativeai>=0.8.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
requests>=2.32.0
html5lib>=1.1
boto3
services
apscheduler>=3.10.4
pytz>=2023.3
